requirements.txt
Flask==1.1.1
Flask-Babel
Flask-Caching
Flask-Cors
#boto3
Pillow
requests
#six==1.12.0
#https://download.pytorch.org/whl/cpu/torch-1.4.0%2Bcpu-cp36-cp36m-linux_x86_64.whl
#https://download.pytorch.org/whl/cpu/torchvision-0.5.0%2Bcpu-cp36-cp36m-linux_x86_64.whl

psutil
gunicorn


#google-cloud-storage